+++ packages/models-library/src/models_library/services_access.py
@@ -5,6 +5,7 @@
 from pydantic import BaseModel, Field
 
 from .users import GroupID
+from .utils.change_case import snake_to_camel
 
 
 class ServiceGroupAccessRights(BaseModel):
@@ -15,6 +16,15 @@ class ServiceGroupAccessRights(BaseModel):
     write_access: bool = Field(
         default=False, description="defines whether the group can modify the service"
     )
+
+
+class ServiceGroupAccessRightsV2(BaseModel):
+    execute: bool = False
+    write: bool = False
+
+    class Config:
+        alias_generator = snake_to_camel
+        allow_population_by_field_name = True
 
 
 class ServiceAccessRights(BaseModel):
